Move octokit types to src/types
