+++ src/octokit/getPullRequest.ts
@@ -1,4 +1,4 @@
-import { ListFilesReturnType, OctokitClient, PullsGetReturnType} from './octokitClient';
+import { ListFilesReturnType, OctokitClient, PullsGetReturnType } from '../types';
 import * as core from '@actions/core';
 import * as github from '@actions/github';
