+++ src/types.ts
@@ -1,3 +1,5 @@
+import * as github from "@actions/github";
+
 export type ConfigFile = {
 	titleTagConfig?: {
 		tags: Array<Tag>
@@ -10,3 +12,9 @@ export type prPatchBody = {
 }
 
 export type Tag = Record<string, string | Array<string>>;
+
+// Octokit Types
+export type OctokitClient = ReturnType<typeof github.getOctokit>;
+export type PullsGetReturnType = Awaited<ReturnType<OctokitClient['rest']['pulls']['get']>>['data'];
+export type PullsUpdateReturnType = Awaited<ReturnType<OctokitClient['rest']['pulls']['update']>>['data'];
+export type ListFilesReturnType = Awaited<ReturnType<OctokitClient['rest']['pulls']['listFiles']>>['data'];
